+++ Tests/SentryProfilerTests/SentryProfilingPublicAPITests.swift
@@ -7,6 +7,7 @@ import XCTest
 // swiftlint:disable file_length
 class SentryProfilingPublicAPITests: XCTestCase {
     private class Fixture {
+        @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
         let options: Options = {
             let options = Options.noIntegrations()
             options.dsn = TestConstants.dsnAsString(username: "SentrySDKTests")
@@ -35,7 +36,9 @@ class SentryProfilingPublicAPITests: XCTestCase {
 
         let currentDate = TestCurrentDateProvider()
         lazy var timerFactory = TestSentryNSTimerFactory(currentDateProvider: currentDate)
+        @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
         lazy var client = TestClient(options: options)!
+        @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
         lazy var hub = SentryHub(client: client, andScope: scope)
     }
 
@@ -52,6 +55,7 @@ class SentryProfilingPublicAPITests: XCTestCase {
         SentryDependencyContainer.sharedInstance().dateProvider = fixture.currentDate
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     override func tearDown() {
         super.tearDown()
 
@@ -668,12 +672,14 @@ extension SentryProfilingPublicAPITests {
 }
 
 private extension SentryProfilingPublicAPITests {
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func givenSdkWithHub() {
         SentrySDKInternal.setCurrentHub(fixture.hub)
         SentrySDKInternal.setStart(with: fixture.options)
         sentry_sdkInitProfilerTasks(fixture.options, fixture.hub)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func givenSdkWithHubButNoClient() {
         SentrySDKInternal.setCurrentHub(SentryHub(client: nil, andScope: nil))
         SentrySDKInternal.setStart(with: fixture.options)