+++ Tests/SentryTests/Integrations/Performance/UIViewController/SentryUIViewControllerPerformanceTrackerTests.swift
@@ -41,6 +41,7 @@ class SentryUIViewControllerPerformanceTrackerTests: XCTestCase {
             return SentryInAppLogic(inAppIncludes: options.inAppIncludes, inAppExcludes: [])
         }
         
+        @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
         init() {
             options = Options.noIntegrations()
             let imageName = String(
@@ -67,6 +68,7 @@ class SentryUIViewControllerPerformanceTrackerTests: XCTestCase {
     
     private var fixture: Fixture!
     
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     override func setUp() {
         super.setUp()
         fixture = Fixture()