+++ Tests/SentryTests/Integrations/Screenshot/SentryScreenshotIntegrationTests.swift
@@ -35,6 +35,7 @@ class SentryScreenshotIntegrationTests: XCTestCase {
         clearTestState()
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func test_attachScreenshot_disabled() {
         SentrySDK.start {
             $0.attachScreenshot = false
@@ -44,6 +45,7 @@ class SentryScreenshotIntegrationTests: XCTestCase {
         XCTAssertFalse(sentrycrash_hasSaveScreenshotCallback())
     }
     
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func test_attachScreenshot_enabled() {
         SentrySDK.start {
             $0.attachScreenshot = true
@@ -53,6 +55,7 @@ class SentryScreenshotIntegrationTests: XCTestCase {
         XCTAssertTrue(sentrycrash_hasSaveScreenshotCallback())
     }
     
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func test_uninstall() {
         SentrySDK.start {
             $0.attachScreenshot = true