+++ Tests/SentryTests/SentrySDKTests.swift
@@ -8,6 +8,7 @@ class SentrySDKTests: XCTestCase {
     
     private class Fixture {
     
+        @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
         let options: Options = {
             let options = Options.noIntegrations()
             options.dsn = SentrySDKTests.dsnAsString
@@ -89,6 +90,7 @@ class SentrySDKTests: XCTestCase {
         fixture = Fixture()
     }
     
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     override func tearDown() {
         super.tearDown()
 
@@ -104,6 +106,7 @@ class SentrySDKTests: XCTestCase {
     }
 
     // Repro for: https://github.com/getsentry/sentry-cocoa/issues/1325
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testStartWithZeroMaxBreadcrumbsOptionsDoesNotCrash() {
         SentrySDK.start { options in
             options.dsn = SentrySDKTests.dsnAsString
@@ -149,6 +152,7 @@ class SentrySDKTests: XCTestCase {
         assertIntegrationsInstalled(integrations: expectedIntegrations)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testStartStopBinaryImageCache() {
         SentrySDK.start { options in
             options.debug = true
@@ -163,6 +167,7 @@ class SentrySDKTests: XCTestCase {
         XCTAssertNil(SentryDependencyContainer.sharedInstance().binaryImageCache.cache)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testStartWithConfigureOptions_NoDsn() throws {
         SentrySDK.start { options in
             options.debug = true
@@ -176,6 +181,7 @@ class SentrySDKTests: XCTestCase {
         XCTAssertEqual(true, options?.debug)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testStartWithConfigureOptions_WrongDsn() throws {
         SentrySDK.start { options in
             options.dsn = "wrong"
@@ -188,6 +194,7 @@ class SentrySDKTests: XCTestCase {
         XCTAssertNil(options?.parsedDsn)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testStartWithConfigureOptions_BeforeSend() {
         var wasBeforeSendCalled = false
         SentrySDK.start { options in
@@ -204,6 +211,7 @@ class SentrySDKTests: XCTestCase {
         XCTAssertTrue(wasBeforeSendCalled, "beforeSend was not called.")
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testStartWithScope() {
         let scope = Scope()
         scope.setUser(User(userId: "me"))
@@ -238,6 +246,7 @@ class SentrySDKTests: XCTestCase {
         XCTAssertFalse(SentrySDK.detectedStartUpCrash)
     }
     
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testInstallIntegrations_NoIntegrations() {
         SentrySDK.start { options in
             options.removeAllIntegrations()
@@ -246,11 +255,13 @@ class SentrySDKTests: XCTestCase {
         assertIntegrationsInstalled(integrations: [])
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testGlobalOptions() {
         SentrySDK.start(options: fixture.options)
         XCTAssertEqual(SentrySDKInternal.options, fixture.options)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testGlobalOptionsForPreview() {
         startprocessInfoWrapperForPreview()
 
@@ -258,6 +269,7 @@ class SentrySDKTests: XCTestCase {
         XCTAssertEqual(SentrySDKInternal.options, fixture.options)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureEvent() {
         givenSdkWithHub()
 
@@ -266,6 +278,7 @@ class SentrySDKTests: XCTestCase {
         assertEventCaptured(expectedScope: fixture.scope)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureEventWithScope() {
         givenSdkWithHub()
 
@@ -275,6 +288,7 @@ class SentrySDKTests: XCTestCase {
         assertEventCaptured(expectedScope: scope)
     }
        
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureEventWithScopeBlock_ScopePassedToHub() {
         givenSdkWithHub()
 
@@ -283,6 +297,7 @@ class SentrySDKTests: XCTestCase {
         assertEventCaptured(expectedScope: fixture.scopeWithBlockApplied)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureEventWithScopeBlock_CreatesNewScope() {
         givenSdkWithHub()
 
@@ -291,6 +306,7 @@ class SentrySDKTests: XCTestCase {
         assertHubScopeNotChanged()
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureError() {
         givenSdkWithHub()
 
@@ -299,6 +315,7 @@ class SentrySDKTests: XCTestCase {
         assertErrorCaptured(expectedScope: fixture.scope)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureErrorWithScope() {
         givenSdkWithHub()
 
@@ -308,6 +325,7 @@ class SentrySDKTests: XCTestCase {
         assertErrorCaptured(expectedScope: scope)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureErrorWithScopeBlock_ScopePassedToHub() {
         givenSdkWithHub()
 
@@ -316,6 +334,7 @@ class SentrySDKTests: XCTestCase {
         assertErrorCaptured(expectedScope: fixture.scopeWithBlockApplied)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureErrorWithScopeBlock_CreatesNewScope() {
         givenSdkWithHub()
 
@@ -324,6 +343,7 @@ class SentrySDKTests: XCTestCase {
         assertHubScopeNotChanged()
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureException() {
         givenSdkWithHub()
 
@@ -332,6 +352,7 @@ class SentrySDKTests: XCTestCase {
         assertExceptionCaptured(expectedScope: fixture.scope)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureExceptionWithScope() {
         givenSdkWithHub()
 
@@ -341,6 +362,7 @@ class SentrySDKTests: XCTestCase {
         assertExceptionCaptured(expectedScope: scope)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureExceptionWithScopeBlock_ScopePassedToHub() {
         givenSdkWithHub()
 
@@ -349,6 +371,7 @@ class SentrySDKTests: XCTestCase {
         assertExceptionCaptured(expectedScope: fixture.scopeWithBlockApplied)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureExceptionWithScopeBlock_CreatesNewScope() {
         givenSdkWithHub()
 
@@ -357,6 +380,7 @@ class SentrySDKTests: XCTestCase {
         assertHubScopeNotChanged()
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureMessageWithScopeBlock_ScopePassedToHub() {
         givenSdkWithHub()
 
@@ -365,6 +389,7 @@ class SentrySDKTests: XCTestCase {
         assertMessageCaptured(expectedScope: fixture.scopeWithBlockApplied)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func testCaptureMessageWithScopeBlock_CreatesNewScope() {
         givenSdkWithHub()
 
@@ -425,11 +450,13 @@ extension SentrySDKTests {
         }
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     private func givenSdkWithHubButNoClient() {
         SentrySDKInternal.setCurrentHub(SentryHub(client: nil, andScope: nil))
         SentrySDKInternal.setStart(with: fixture.options)
     }
 
+    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     private func givenSdkWithHub() {
         SentrySDKInternal.setCurrentHub(fixture.hub)
         SentrySDKInternal.setStart(with: fixture.options)